Совместимость двух компиляторов MSVC и GCC

+++ xface/point.h
@@ -7,8 +7,8 @@ struct point {
 	explicit			operator bool() const { return x || y; }
 	bool				operator!=(const point pt) const { return pt.x != x || pt.y != y; }
 	bool				operator==(const point pt) const { return pt.x == x && pt.y == y; }
-	point				operator-(const point pt) const { return{x - pt.x, y - pt.y}; }
-	point				operator+(const point pt) const { return{x + pt.x, y + pt.y}; }
+	point				operator-(const point pt) const { return{(short)(x - pt.x), (short)(y - pt.y)}; }
+	point				operator+(const point pt) const { return{(short)(x + pt.x), (short)(y + pt.y)}; }
 	void				clear() { x = y = 0; }
 	bool				in(const rect& rc) const { return x >= rc.x1 && x <= rc.x2 && y >= rc.y1 && y <= rc.y2; }
 	bool				in(const point p1, const point p2, const point p3) const;
